backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
apscheduler==3.10.4
openpyxl==3.1.5
pandas==2.2.3
bcrypt==4.2.0
pyjwt==2.9.0
uiautomation==2.0.20
pyinstaller==6.11.1
chinesecalendar==1.10.0
